requirements.txt
beautifulsoup4==4.4.1
Django==1.9.8
django-filter==0.13.0
django-object-actions==0.8.2
django-redis-cache==1.6.5
djangorestframework==3.3.3
drf-nested-routers==0.11.1
eventlet==0.19.0
gevent==1.1.1
greenlet==0.4.10
gunicorn==19.6.0
html2text==2016.5.29
Jinja2==2.8
jsonfield==1.0.3
MarkupSafe==0.23
psycopg2==2.6.1
py==1.4.31
pytest==2.9.2
pytest-django==2.9.1
qiniu==7.0.6
redis==2.10.5
requests==2.10.0
rules==1.1.1
six==1.10.0
uWSGI==2.0.13.1
